+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes
 0.3.1 (unreleased)
 
 - Prevent CPD errors when SYNTAX_ERROR token is too long
+- Fix Magik Grammar bug not being able to parse expressions and parenthesis
 
 
 0.3.0 (2019-09-22)